refactor(frontend): extract shared CardListScreen for cards/master-cards clients
